sched/uclamp: Use wrappers for *_{read, write, show} functions
In commit:
    sched/uclamp: Move all tunables to cpusets,

We directly modified and exported *_{read, write, show} functions
to cpuset code.This makes the resulting code quite inconsistent with
the upstream version.

We now use wrappers to reflect the original function, and to preserve
the code in accordance with upstream/mainline.

This is a purely cosmetic change, and no new behaviour should be
exhibited on applying this commit.

Change-Id: I8e96c4ea37af3a2cbf2fb7a4d0a2ea25d5e0230f
Signed-off-by: strongreasons <student-jmxdeqtpi@student.sman1sumenep.sch.id>